refactor: move xnvme_opts_default() from xnvme_be.c to xnvme_opts.c
